# docs/openapi/openapi.yaml
...
openapi: 3.0.4
...
components:
  schemas:
...
    ChunkedUploadInitRequest:
      type: object
      properties:
        fileName:
          type: string
        fileSize:
          format: int64
          type: integer
        importFormat:
          type: string
        options:
          type: string
...
    ConnectionType:
      enum:
      - HTTP
      - MQTT
      type: string
...
    CreateGpsSourceConfigDto:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/GpsSourceType"
        username:
          type: string
        password:
          type: string
        token:
          type: string
        deviceId:
          type: string
        userId:
          $ref: "#/components/schemas/UUID"
        connectionType:
          $ref: "#/components/schemas/ConnectionType"
        filterInaccurateData:
          type: boolean
        maxAllowedAccuracy:
          format: int32
          type: integer
        maxAllowedSpeed:
          format: int32
          type: integer
        enableDuplicateDetection:
          type: boolean
        duplicateDetectionThresholdMinutes:
          format: int32
          type: integer
...
    GpsSourceType:
      enum:
      - OWNTRACKS
      - GPSLOGGER
      - OVERLAND
      - TRACCAR
      - GOOGLE_TIMELINE
      - GPX
      - DAWARICH
      - HOME_ASSISTANT
      - GEOJSON
      - CSV
      - COLOTA
      - MANUAL
      - MOBILE_APP
      type: string
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
    UpdateGpsSourceConfigDto:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        username:
          type: string
        password:
          type: string
        token:
          type: string
        deviceId:
          type: string
        userId:
          type: string
        connectionType:
          $ref: "#/components/schemas/ConnectionType"
        filterInaccurateData:
          type: boolean
        maxAllowedAccuracy:
          format: int32
          type: integer
        maxAllowedSpeed:
          format: int32
          type: integer
        enableDuplicateDetection:
          type: boolean
        duplicateDetectionThresholdMinutes:
          format: int32
          type: integer
    UpdateGpsSourceConfigStatusDto:
      type: object
      properties:
        status:
          type: boolean
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /api/gps/source:
    put:
      operationId: GpsSourceConfigResource_updateGpsConfigSource
      tags:
      - "User: GPS Sources"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGpsSourceConfigDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Update Gps Config Source
      security:
      - jwt:
        - USER
        - ADMIN
    get:
      operationId: GpsSourceConfigResource_getGpsSourceConfigs
      tags:
      - "User: GPS Sources"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Gps Source Configs
      security:
      - jwt:
        - USER
        - ADMIN
    post:
      operationId: GpsSourceConfigResource_addGpsSourceCnfig
      tags:
      - "User: GPS Sources"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGpsSourceConfigDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Add Gps Source Cnfig
      security:
      - jwt:
        - USER
        - ADMIN
...
  /api/gps/source/{id}:
    delete:
      operationId: GpsSourceConfigResource_deleteGpsSourceConfig
      tags:
      - "User: GPS Sources"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Delete Gps Source Config
      security:
      - jwt:
        - USER
        - ADMIN
  /api/gps/source/{id}/status:
    put:
      operationId: GpsSourceConfigResource_updateStatus
      tags:
      - "User: GPS Sources"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGpsSourceConfigStatusDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Update Status
      security:
      - jwt:
        - USER
        - ADMIN
...
  /api/import/upload:
    post:
      operationId: UnifiedImportResource_uploadFile
      tags:
      - "User: Import and Export"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  format: binary
                  type: string
                format:
                  type: string
                options:
                  type: string
            encoding:
              options:
                contentType: text/plain
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Upload File
      security:
      - jwt: []
  /api/import/upload/init:
    post:
      operationId: UnifiedImportResource_initializeChunkedUpload
      tags:
      - "User: Import and Export"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChunkedUploadInitRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
        "400":
          description: Bad Request
      summary: Initialize Chunked Upload
      security:
      - jwt: []
  /api/import/upload/{uploadId}:
    delete:
      operationId: UnifiedImportResource_abortUpload
      tags:
      - "User: Import and Export"
      parameters:
      - name: uploadId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Abort Upload
      security:
      - jwt: []
  /api/import/upload/{uploadId}/chunk:
    post:
      operationId: UnifiedImportResource_uploadChunk
      tags:
      - "User: Import and Export"
      parameters:
      - name: uploadId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                chunkIndex:
                  format: int32
                  type: integer
                chunk:
                  format: binary
                  type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Upload Chunk
      security:
      - jwt: []
  /api/import/upload/{uploadId}/complete:
    post:
      operationId: UnifiedImportResource_completeChunkedUpload
      tags:
      - "User: Import and Export"
      parameters:
      - name: uploadId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Complete Chunked Upload
      security:
      - jwt: []
  /api/import/upload/{uploadId}/status:
    get:
      operationId: UnifiedImportResource_getUploadStatus
      tags:
      - "User: Import and Export"
      parameters:
      - name: uploadId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Authentication is required or the provided credentials are
            invalid.
        "403":
          description: The authenticated user does not have permission to access this
            resource.
      summary: Get Upload Status
      security:
      - jwt: []
...
info:
  title: GeoPulse API
  version: 1.34.1
  description: GeoPulse backend REST API documentation.
  contact:
    name: tess1o
    email: kerriden1@gmail.com
servers:
- url: http://localhost:8080